rework encoded plugin to include PolygonFromEncoded
